WIP: rename service to servicedeployer
