+++ internal/testrunner/runners/system/servicedeployer/compose.go
@@ -1,4 +1,4 @@
-package service
+package servicedeployer
 
 import (
 	"fmt"
