+++ internal/testrunner/runners/system/servicedeployer/factory.go
@@ -1,4 +1,4 @@
-package service
+package servicedeployer
 
 import (
 	"errors"
